requirements.txt
streamlit
requests
python-dotenv
langchain
langchain-openai
langchain-community